Fix a latent bug in print_wchar
print_wchar keeps track of when escape sequences are emitted, to force
an escape sequence if needed by a subsequent character.  For example
for the string concatenation "\0" "1", gdb will print "\000\061" --
because printing "\0001" might be confusing.

However, this code has a logic error.  It's intended to reject '8' and
'9', as they are not octal digits, but it fails to do so.  This patch
fixes the bug, and slightly rewrites the expression to be a bit more
clear.

I suspect that, actually, this code can be removed entirely, because
octal sequences are limited to 3 digits.  However, I'm not certain, so
I've left it in place.