Add an emitter callback to generic_printstr and generic_emit_char
This adds an emitter callback to generic_printstr and
generic_emit_char, passing it through to print_wchar.  print_wchar is
modified to call it, if possible.  This will be used to let languages
override the way that escape sequences are emitted.  Nothing uses this
yet, that comes later in the series.