Add an emitter callback to generic_printstr and generic_emit_char
This adds an emitter callback to generic_printstr and
generic_emit_char.  print_wchar is renamed, exported, and modified to
be suitable for use as the default emitter.  This will be used to let
languages override the way that escape sequences are emitted.  Nothing
uses this yet, that comes later in the series.