Don't use realpath on objfile names
Currently gdb uses realpath when finding an objfile.  However, this
yields surprising results when the given name is a symlink -- if the
symlink is retargeted, re-running will not cause a change.

This patch fixes the problem by changing gdb to compute the absolute
path rather than the real path.

This area has some history.  PR gdb/15415 changed this area to try to
preserve the basename of the executable, so that programs that check
argv[0] will continue to work correctly.

There's also PR gdb/15934, which proposes preserving the original name
for use in argv[0] and in "info inferior".  I haven't done this.  The
"info inferior" part might not be too difficult (unsure) but I wasn't
sure how argv[0] could really be preserved given relative file names,
the use of shell to startup, and the user's ability to change gdb's
working directory.

I think this patch does fix PR symtab/24143, which concerns finding
separate debug files when symlinks are used.  The included test does
not test exactly this scenario, though.

Regression tested on x86-64 Fedora 34.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=24143