Pre-map all debug sections
This changes the DWARF reader to pre-read all the section data.  This
is how gdb used to work ages ago, but I changed it to be lazier.  That
turns out to have been a bit of a bug -- PR gdb/14202 points out that
this fails if the .so is unlinked after the program starts.

The laziness doesn't seem to matter so much any more, and this
simplifies the code as well.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=14202