Only allow downcasts in checked_static_cast
It didn't make sense to me that checked_static_cast allows upcasts, as
it seems to me that it is better to simply not write an explicit
upcast, but instead just allow the implicit derived-to-base
conversion.

This patch removes the downcast check from checked_static_cast.  It
also doesn't allow casts to the identical type, as this also did not
seem necessary.

Tested by rebuilding.