Use Tcl auto-loading
While working on the earlier patches, I noticed there's some
inconsistency in whether "load_lib" or "require" should come first in
the .exp files.  The answer is that load_lib should come first -- but
Tcl has had an auto-loading feature for a very long time, and it
seemed better to me to simply use that.  This way, the precise
location of procs does not need to be encoded into every test file.
Instead, procs are simply loaded on demand.

When updating a file in testsuite/lib, the auto-load index must be
refreshed using the make-tcl-index script.  This should not be
difficult to remember, because tests will not work without it.