Introduce the "critical" testsuite command
This patch introduces a new "critical" command for the test suite.
Unlike require, "critical" can handle multiple arguments and a variety
of return value conventions.  It is implemented as a Tcl namespace
ensemble for this reason.

This version of the patch provides a few such subcommands.  Currently
these are all implemented via a helper proc, but there's no
requirement that this always be true.