Simplify cp_print_class_member
I noticed that cp_print_class_member's calling convention can be
simplified.  In particular it seems cleaner for it to simply take a
value and a stream; and the "prefix" argument is only ever set to one
value.

This version also renames the function, to indicate a bit more clearly
what it actually does.

Regression tested on x86-64 Fedora 40.

Reviewed-By: Keith Seitz <keiths@redhat.com>