Add minimal thread-safety to BFD
This patch provides some minimal thread-safety to BFD.

The globals used during BFD creation (e.g., bfd_id_counter) are
wrapped in locks, and the file descriptor cache is also locked.  A
function to clean up any thread-local data is now provided for BFD
clients.

Currently the only threading primitive that is needed is a mutex.
POSIX and Windows ports are provided.