Return a block_symbol from find_pc_sect_function
This changes find_pc_sect_function to return a block_symbol and
updates all the callers in a minimal way.

A less intrusive approach might be possible -- but in the long run
we'll want to add an objfile into this result, so it makes sense to go
ahead with the conversion now.