Add block_symbol::address
This adds a new block_symbol::address method.  The long-term goal here
is to change most existing calls to symbol::value_address to call
block_symbol::address instead.

Then, at some future date, we can add an objfile to block_symbol and
change this method to relocate the address when called -- making a
symbol independent of the objfile, as we've already done for minimal
symbols and partial symbols.