Introduce block_symbol::objfile
When we finally managed to make blocks independent of the progspace,
block_symbol (and other "bound" wrappers) will hold a pointer to the
objfile.  Meanwhile, we can prepare the users for this change by
introducing a method to fetch the objfile.  This patch does this.
Right now the method just forwards to the block; but in the future it
can be changed to return the member instead.