Introduce and use compunit_symtab::arch
Eventually, we'd like to remove objfile back-links from
compunit_symtab.  To help prepare for this change, this patch
introduces a new compunit_symtab::arch method.  Right now this just
forwards the call via the objfile, but in the future this can be
changed to use the BFD instead, without requiring updates to all the
callers.  This kind of change also makes it simpler to find the uses
of the backlinks.