Compare section index in lookup_minimal_symbol_by_pc_section
This changes lookup_minimal_symbol_by_pc_section to compare the
section index when comparing two symbols.  This is ok because the
symbols in question always come from the same objfile.  The motivation
for this is to simplify a refactoring in a subsequent patch; but it is
also more efficient.