Hoist a call to frob_address
This hoists the call to frob_address in
lookup_minimal_symbol_by_pc_section, so that it now appears earlier in
the loop.  This is useful to help with a refactoring in the next
patch.  Making this change also made it clear that 'pc' is not used,
so this removes the local variable as well.