Add unrelocated overload of lookup_minimal_symbol_by_pc_section
This refactors lookup_minimal_symbol_by_pc_section, pulling out the
inner loop into a separate function.  This function also does all its
work using only the per-BFD object -- not the objfile.

This overload is used in the next patch, but it seemed simpler to
review as its own patch.