Normalize "on|off" text in manual
While looking in the info manual, I noticed this:

    ‘set style enabled ‘on|off’’

The styling looked weird here and in the source I saw it was written as:

    @item set style enabled @samp{on|off}

I looked through the rest of the manual and found this same sort of
thing was written in a variety of styles.

For this patch, I picked one way to write this and applied it to all
the instances.