Add check-gdbarch.py
This adds a new check-gdbarch.py script.  This script checks the
sources to see which gdbarch methods are set but never called, and
which ones are called but never set.

This uses Pygments to scan the sources, to ensure we don't
accidentally pick up gdbarch references from comments (or, less
likely, strings).  Because this is pretty slow, it also uses "tqdm" to
show a progress bar.  (You can "pip install tqdm" to get this.)

This version also sorts the output.