Implement local-environment commands
A user noted that "set environment" does not affect things like
"shell" or "pipe".  This makes some sense, because the environment is
per-inferior, and also in the general case it may be affecting a
remote machine -- i.e., the settings may not be locally appropriate.

This patch adds a new set of "local-environment" commands (set, show,
and unset) that affect gdb's own environment.