Require current language before symbol lookups
py-symbol.exp fails with various target boards, including fission and
gold-gdb-index.

The problem here is that, in this test, the current language is still
unset (i.e., lazy) when the symbol lookup is done.  It is eventually
set deep in the lookup -- but this then requires a reentrant symbol
lookup, which fails.  (DWARF symbol lookup is not reentrant.)

This patch changes the top-level symbol-lookup functions to require
that the current language be available.  This fixes the bug.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=32490