Mention gdb thread ID in new thread message
Currently the "new thread" message that gdb might print does not
include gdb's own thread ID -- so, if you want to reference the
thread, you must first use "thread find" or "info threads".

This patch changes the announcement to mention the thread.  It's
mildly ugly because the thread-to-string code adds a redundant
"Thread" string.

I wasn't sure whether this code should use print_full_thread_id or
just print_thread_id.  I went with the former, on the theory that if
new inferior is started as well, it's handy to look back and see the
full ID; but this is easily changed if desirable.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=19584