Put all CTF symbols in global scope
The new approach to searching (solely via the quick API) is more
sensitive to discrepancies between the partial and full readers.  In
CTF, there is some disagreement about which scope to use.  CTF doesn't
seem to really distinguish between the file and global scope, so this
patch takes the simple approach of putting all CTF symbols into the
global scope.

This changes one test as well.  It seems to me that the behavior here
is arbitrary and the test is making unwarranted assumptions.