Track parents differently in the .debug_names reader
The .debug_names reader implements its own scheme to connect entries
to their parents.  There's nothing wrong with this.  However, this
requires some extra work when parallelizing the reader, and since the
cooked indexer already solves this problem -- a read-only map that's
consulted during finalization, which is threaded -- it seemed better
to switch .debug_names to use this same scheme.

Due to the support for old "GDB2" indexes, and because the parent IDs
don't really refer to DIEs, the .debug_names reader implements its own
scheme for computing parent_map::addr_type values.  These values are
essentially meaningless tokens, though, so this seems fine.

I tried using a parent_map here, but this resulted in a substantial
slowdown.  And, because .debug_names does not need to track DIE ranges
-- parents are always referenced by their specific number -- this
patch instead adds hash tables as a secondary lookup in
parent_map_map.