Parallelize the .debug_names reader FIXME - measure
This parallelizes the .debug_names reader.  The general approach is
somewhat similar to the cooked index reader.  Perhaps a bit more code
could even be shared here.

On my machine (8 cores), this improves performance a little, but not
very much.  When run on a fixed copy of gdb with .debug_names,
performance goes from:

4.95user 0.34system 0:01.82elapsed 290%CPU (0avgtext+0avgdata 602264maxresident)k

to

3.41user 0.31system 0:01.72elapsed 216%CPU (0avgtext+0avgdata 519948maxresident)k