gnu directives: gas/readelf tests for gnu attributes v2
These tests are a copy-paste of the generic parsing tests for AArch64.

The added tests cover the parsing of the new assembly directives
(gnu_subsection and gnu_attribute), the serialization of the
Object Attributes v2 (OAv2) data into an object file, and the
dumping of those data via readelf.

The parsing tests focus on the following points:
- the syntax of the new directives.
- the recognition of generic tokens like: NTBS, ULEB128, required,
  optional.

The dumping tests focus on:
- the OAv2 population into the correct section assigned by the backend.
- the merge of the subsections and attributes when they are declared
  several times inside respectively the same compilation unit, and
  subsection.
- the sorting of OAv2 before the serialization.