gdb: create address map after parsing all DIE
Continuing the work done in the last two commits, this commit defers
building the addrmap for a blockvector until after all the DIE have
been read, and the line table processed.

The benefit of this is that any changes to a block's ranges done
during line table processing (see the next commit) will be reflected
in the blockvector's addrmap.

The alternative to this is to build the addrmap as we initially see
each block, but then adjust the addrmap if we later decide to modify a
block.  I think deferring the addrmap creation is cleaner, and is less
work overall.

The addrmap requires that we add the innermost blocks first.  I
achieve this by walking the blockvector backward, as we always add
parent blocks before their more inner child blocks.

However, the blocks are not added to the blockvector in exactly the
same order as we recorded them before this patch.  The key difference
here is that sibling blocks might not appear in the block vector in
the same order as we parsed them from the DWARF.

So long as sibling blocks don't overlap, this shouldn't make any
difference to the final addrmap that is constructed.  But if the
sibling blocks do overlap, then there might be an observable change in
behaviour after this patch.  But it is not clear to me how critical
this change actually is.  As far as I can tell from my experiments, if
two siblings overlap then, for the overlap region, GDB is only able to
find information (e.g. local variables) from one sibling.  This is
true before and after this patch, but which sibling is found might
change.

Approved-By: Tom Tromey <tom@tromey.com>