Change how C parser looks up type tag names
In an earlier version of this series, Simon noticed that something
like "ptype struct type" was much slower.

The deep problem here is the hack in best_symbol -- in this particular
case the C parser is including SEARCH_VAR_DOMAIN, which means that a
type is not "best".

Fixing best_symbol would be good to do, but it's difficult and I
already had many struggles getting this series to this point.  So,
rather than fix that, I elected to modify the parser.

The key insight here is that the name lookup in classify_name is
useless when a tag name is expected.  This patch adds a flag so that
this lookup is not done.  (Incidentally, this kind of thing is much
more straightforward in a recursive descent parser, something like
classify_name would only be applied post-lexing where appropriate.)

This change speeds up the lookup by avoiding the best_symbol hack,
instead searching only for a type.