Add evaluateName to registers returned by DAP
This patch adds the evaluateName field to register variables returned
by DAP.

This isn't truly complete, in that referring to a sub-part of a
register won't have an evaluateName.  This is a larger problem,
though, and requires a solution for pretty-printers as well.

Meanwhile, this patch helps with a pretty typical case.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=33407