Add quotemeta to gdb test suite
AdaCore has an internal test suite written in Python.  One nice
facility it has is called "quotemeta", and it is an alternative to
using regular expressions to match gdb output.

This patch adds a similar library, written in Tcl, to the gdb test
suite.  The idea here is to simplify certain kinds of checks and to
avoid the painful combination of both Tcl and regexp quoting.

There is a comment in the code that explains quotemeta in detail.  If
you've used the AdaCore test suite, note that this implementation has
some extensions, namely the @{...} forms.  (I do wonder here if <...>
would be a better choice, to avoid Tcl issues with braces.)

I've converted part of a Rust test case to demonstrate the effect.
The most dramatic example is this change:

-gdb_test "print g" " = \\(\\*mut \\\[u8; 6\\\]\\) $hex b\"hi bob\""
+gdb_test "print g" [quotemeta {$@DECIMAL = (*mut [u8; 6]) @HEX b"hi bob"}]

At least for me the latter is much easier to understand; and also it
doesn't cheat on the history part of the output, because it's less
painful to do this properly now.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=33551
Reviewed-By: Tom de Vries <tdevries@suse.de>