Avoid crashes when stepping through green thread context-switching
When stepping through the green thread context-switching code, gdb
will try to single-step the current thread, while leaving the other
threads stopped.  This results in an assertion failure in
finish_step_over, because the event thread is not the request thread.

This patch introduces a new target method so that infrun can ask the
target whether this can happen; and then changes infrun to handle this
case.  When the target can have an unexpected thread switch,
single-stepping may also switch threads.