gdb/testsuite: add test for corefile with no threads
This commit adds a test for the recent commit:

  commit af7fe6fff91a61f30f9adddc7ff2f8852dc6482a
  Date:   Fri Nov 21 14:38:01 2025 +0100

      [gdb/corefiles] Fix segfault in add_thread_silent

The previous commit fixed an issue where BFD was failing to parse the
NT_PRSTATUS notes in a core file.  As a result of this the general
purpose registers were not available to GDB, and as a result GDB would
not add any threads to the inferior when opening a core file.

As GDB requires each inferior to have at least one thread, there is
code in GDB to handle this no-registers case, and add a dummy thread
to the inferior.  However, a bug in this code was causing GDB to
crash.  It is this crash that the above commit fixed.

This commit adds a test for this fix.

The test contains a Python script which adds a new Python command to
GDB.  This Python command can be used to modify an on-disk core file.
The command opens the core file, locates the notes segments, and then
looks through the notes to find NT_PRSTATUS notes.  The type value for
these NT_PRSTATUS notes is then changed to a value that GDB doesn't
understand (0xffffffff).

Now when GDB opens the core file the NT_PRSTATUS are no longer
NT_PRSTATUS notes; the core file will appear to have no such notes,
and the previous bug will be triggered.

Running this new test with GDB 16 will trigger the crash.  With
current HEAD of master, which includes the above commit, this test
should pass.

I originally wrote the Python script in this test as a standalone
Python script, not as a GDB command, but some of the gcc compiler farm
test machines still have Python 2 as the default.  This means that the
corefile-no-threads.exp script would have to figure out which Python
executable to use to run the script.

Much easier to just use GDB's builtin Python interpreter.  We know
that this will be at least Python 3.4 (GDB's minimum Python
requirement).

I've tested this on x86-64 in 64-bit and 32-bit mode.  And on a big
endian PPC64 machine.