Restore ui_file::can_page
A while back, I removed the ui_file::can_page method.  It wasn't
needed for a while, due to how output redirections were done.

With the new approach in this series, it's convenient to have this
method again, because the pipeline downstream from the pager won't
change when logging -- the logging filter will be enabled or disabled
without reconfiguring the pipeline.  This method lets the pager adapt
correctly.

Approved-By: Andrew Burgess <aburgess@redhat.com>