Fix DAP 'disconnect' implementation
gdb's implementation of the DAP 'disconnect' request was incorrect in
a few ways.

First, the 'terminateDebuggee' field is optional, and has a special
meaning when not supplied: it should do whatever the default is.

Second, if the inferior was attached, it should detach rather than
terminate by default.

Finally, if the inferior was not started at all, it seems reasonable
for this request to simply succeed silently -- currently it returns
"success: false" with the reason being that the inferior isn't
running.

Approved-By: Andrew Burgess <aburgess@redhat.com>